flytter sletting til /draft/id

+++ src/main/kotlin/no/nav/hm/grunndata/register/series/SeriesRegistrationCommonController.kt
@@ -267,6 +267,23 @@ class SeriesRegistrationCommonController(
             if (seriesToUpdate.published != null) throw BadRequestException("can not delete a published series")
         }
 
+        seriesRegistrationService.deleteSeries(seriesToUpdate, authentication)
+        LOG.info("set series to deleted: $id")
+        return HttpResponse.ok()
+    }
+
+    @Delete("/draft/{id}")
+    suspend fun deleteDraftSeries(
+        @PathVariable id: UUID,
+        authentication: Authentication,
+    ): HttpResponse<Any> {
+        val seriesToUpdate = seriesRegistrationService.findById(id, authentication) ?: return HttpResponse.notFound()
+
+        if (authentication.isSupplier()) {
+            if (seriesToUpdate.draftStatus != DraftStatus.DRAFT) throw BadRequestException("series is not a draft")
+            if (seriesToUpdate.published != null) throw BadRequestException("can not delete a published series")
+        }
+
         seriesRegistrationService.deleteDraft(seriesToUpdate, authentication)
         LOG.info("set series to deleted: $id")
         return HttpResponse.ok()
